+++ mods/BrewLAN_Units/BrewAir/lua/CustomUnits/T2Fighters.lua
@@ -3,22 +3,22 @@ UnitList = {
     -- T2 Bombers
     ----------------------------------------------------------------------------
     T1AirFighter = {
-        UEF = {'sea0202', 40},
+        UEF = {'sea0212', 40},
         --Cybran = {'sra0202', 40},
         --Seraphim = {'ssa0202', 40},
     },
     T2FighterBomber = {
-        UEF = {'sea0202', 30},
+        UEF = {'sea0212', 30},
         --Cybran = {'sra0202', 30},
         --Seraphim = {'ssa0202', 30},
     },
     T2AirFighter = {
-        UEF = {'sea0202', 90},
+        UEF = {'sea0212', 90},
         --Cybran = {'sra0202', 90},
         --Seraphim = {'ssa0202', 90},
     },
     T3AirFighter = {
-        UEF = {'sea0202', 20},
+        UEF = {'sea0212', 20},
         --Cybran = {'sra0202', 20},
         --Seraphim = {'ssa0202', 20},
     },