+++ mods/BrewLAN_Units/BrewAir/units/SEA0212/SEA0212_Script.lua
@@ -3,11 +3,11 @@ local TWeapons = import('/lua/terranweapons.lua')
 local TAAGinsuRapidPulseWeapon = TWeapons.TAAGinsuRapidPulseWeapon
 local TIFCruiseMissileLauncher = TWeapons.TIFCruiseMissileLauncher
 
-SEA0202 = Class(TAirUnit) {
+SEA0212 = Class(TAirUnit) {
     Weapons = {
         AutoCannon = Class(TAAGinsuRapidPulseWeapon) {},
         Missile = Class(TIFCruiseMissileLauncher) {},
     },
 }
 
-TypeClass = SEA0202
+TypeClass = SEA0212