+++ mods/BrewLAN_Units/BrewResearch/hook/lua/system/Blueprints.lua
@@ -16,6 +16,23 @@ function ModBlueprints(all_blueprints)
     GenerateResearchItemBPs(all_blueprints.Unit)
 end
 
+function WikiBlueprints(all_blueprints)
+    for id, bp in pairs(all_blueprints.Unit) do
+        local t = table.find(bp.Categories, 'TECH3') and 3 or table.find(bp.Categories, 'TECH2') and 2 or table.find(bp.Categories, 'TECH1') and 1
+        if TableFindSubstrings(bp.Categories, 'BUILTBY', 'FACTORY') and bp.Physics then
+            if bp.Physics.MotionType == 'RULEUMT_Hover' or bp.Physics.MotionType == 'RULEUMT_AmphibiousFloating' then
+                for i = t, 3 do
+                    table.insert(bp.Categories, 'BUILTBYTIER'..i..'SURFACEFACTORY')
+                end
+            elseif bp.Physics.MotionType == 'RULEUMT_Amphibious' then
+                for i = t, 3 do
+                    table.insert(bp.Categories, 'BUILTBYTIER'..i..'SEABEDFACTORY')
+                end
+            end
+        end
+    end
+end
+
 --------------------------------------------------------------------------------
 -- Things in preparation of RND
 --------------------------------------------------------------------------------
@@ -74,66 +91,6 @@ function RestrictExistingBlueprints(all_bps)
         end
     end
 end
-
---------------------------------------------------------------------------------
--- Rebalance a few vanilla units
---------------------------------------------------------------------------------
---[[function RebalanceExistingBlueprints(all_bps)
-    local t3radars = {
-        uab3104 = 'sab3301',
-        ueb3104 = 'seb3301',
-        urb3104 = 'srb3301',
-        xsb3104 = 'ssb3302',--Forgot the optics tracking facility was already SSB3301
-    }
-    for id, omniID in t3radars do
-        local bp = all_bps[id]
-        local omnibp = all_bps[omniID]
-        if bp and omnibp then
-            if bp.Intel.OmniRadius and bp.Intel.OmniRadius ~= 0 and bp.Economy.MaintenanceConsumptionPerSecondEnergy and bp.Economy.BuildCostEnergy and bp.Economy.BuildCostMass and bp.Economy.BuildTime then
-                --Scale omni radii
-                omnibp.Intel.OmniRadius = bp.Intel.OmniRadius * 1.5
-                bp.Intel.OmniRadius = nil
-
-                --Remove omni categories
-                if bp.Categories and type(bp.Categories) == 'table' then
-                    table.removeByValue(bp.Categories, 'OMNI')
-                    table.removeByValue(bp.Categories, 'OVERLAYOMNI')
-                    if not table.find(bp.Categories, 'RADAR') then
-                        table.insert(bp.Categories, 'RADAR')
-                    end
-                end
-
-                --Remove omni visiual elements
-                if bp.Display.Abilities and type(bp.Display.Abilities) == 'table' then
-                    table.removeByValue(bp.Display.Abilities, 'Omni')
-                    table.removeByValue(bp.Display.Abilities, 'Omni Sensor')
-                    table.removeByValue(bp.Display.Abilities, '<LOC ability_omni>Omni Sensor')
-                end
-                if bp.General.OrderOverrides.RULEUTC_IntelToggle then
-                    bp.General.OrderOverrides.RULEUTC_IntelToggle.bitmapId = 'radar'
-                    bp.General.OrderOverrides.RULEUTC_IntelToggle.helpText = 'toggle_radar'
-                end
-                bp.Description = '<LOC ueb3201_desc>Radar System'
-
-                --Scale health
-                omnibp.Defense.Health = bp.Defense.Health * 5
-                omnibp.Defense.MaxHealth = bp.Defense.MaxHealth * 5
-
-                --Scale costs
-                omnibp.Economy.MaintenanceConsumptionPerSecondEnergy = bp.Economy.MaintenanceConsumptionPerSecondEnergy * 1.5
-                omnibp.Economy.BuildCostEnergy = bp.Economy.BuildCostEnergy * 1.2
-                omnibp.Economy.BuildCostEnergy = bp.Economy.BuildCostEnergy * 1.2
-                omnibp.Economy.BuildTime = bp.Economy.BuildTime * 1.2
-
-                --Adjust costs
-                bp.Economy.MaintenanceConsumptionPerSecondEnergy = bp.Economy.MaintenanceConsumptionPerSecondEnergy * 0.5
-                bp.Economy.BuildCostEnergy = bp.Economy.BuildCostEnergy / 3 * 2
-                bp.Economy.BuildCostMass = bp.Economy.BuildCostMass / 3 * 2
-                bp.Economy.BuildTime = bp.Economy.BuildTime / 3 * 2
-            end
-        end
-    end
-end]]
 
 --------------------------------------------------------------------------------
 -- Create build categories for the amphib/sub/seaplane factories
@@ -371,7 +328,7 @@ end
 
 function TableFindSubstrings(array, string1, string2)
     if array then
-        for i, cat in array do
+        for i, cat in ipairs(array) do
             if string.find(cat,string1) and string.find(cat,string2 or string1) then
                 return i
             end