+++ mods/BrewLAN_Units/BrewTeaParty/hook/lua/system/Blueprints.lua
@@ -13,12 +13,16 @@ function ModBlueprints(all_blueprints)
     BrewLANPersonalShieldCategory(all_blueprints.Unit)
 end
 
+function WikiBlueprints(all_blueprints)
+    BrewLANPersonalShieldCategory(all_blueprints.Unit)
+end
+
 --------------------------------------------------------------------------------
--- This fixes GAZ_UI asuming that just because it has a shield that the unit knows how good it is.
+-- Lets anti-shield units target personal shield units automatically.
 --------------------------------------------------------------------------------
 
 function BrewLANPersonalShieldCategory(all_bps)
-    for id, bp in all_bps do
+    for id, bp in pairs(all_bps) do
         if bp.Categories and bp.Defense and bp.Defense.Shield then
             if (bp.Defense.Shield.PersonalShield or math.max(bp.Defense.Shield.ShieldSize, bp.Defense.Shield.ShieldProjectionRadius or 0) < math.max(bp.SizeX or 1, bp.SizeY or 1, bp.SizeZ or 1) + 3) and table.find(bp.Categories, 'SHIELD') then
                 table.removeByValue(bp.Categories, 'SHIELD')
@@ -28,7 +32,7 @@ function BrewLANPersonalShieldCategory(all_bps)
             end
         end
         if bp.Weapon then
-            for i, wep in bp.Weapon do
+            for i, wep in ipairs(bp.Weapon) do
                 if wep.TargetPriorities and wep.DamageToShields and wep.DamageToShields > (wep.Damage or 0) * 10 then
                     if wep.DamageToShields > (wep.Damage or 0) * 100 then
                         wep.TargetPriorities = {